fix(skynet-srv): half-close forwardRawTCP before close — drains buffered bytes (#2731) (#2735)
* fix(skynet-srv): half-close forwardRawTCP destination before full close — drains buffered bytes (#2731)

Closes #2731. Symptom: first connection through a freshly-established
skynet route truncated at exactly 65536 bytes; subsequent transfers
partial-RST'd at varying offsets (4 MB, 7 MB, etc.). Direct loopback
to the same target was byte-clean. 50 MB cli dmsg cat sends over
--transport skynet RST'd on every attempt.

Root cause: forwardRawTCP's "close both conns when first io.Copy
returns" pattern. io.Copy returns when bytes are accepted by the
destination's SEND BUFFER (smux MaxStreamBuffer = 64 KB), not when
delivered to the peer. When the local-app finishes writing
(memory-fast), local->remote io.Copy returns with all 10 MB
"sent" — but only ~64 KB has reached the actual peer. The
immediate Close(remoteConn) then drops the still-buffered bytes.

The 65536 figure was the smoking gun: exactly one smux
MaxStreamBuffer worth of bytes in flight at close time.

Fix: half-close-on-first-done. When one io.Copy returns:
- Call CloseWrite on the OTHER side's destination (when supported).
  net.TCPConn, hashicorp/yamux.Stream, and xtaci/smux.Stream all
  expose CloseWrite() — half-close signals EOF to the peer without
  killing the receive direction, letting the send buffer drain.
- Wait (bounded by forwardRawTCPDrainTimeout = 30s) for the second
  direction's io.Copy to finish naturally.
- THEN fully close both conns.

Falls back to immediate Close for conn types without CloseWrite —
no worse than pre-fix behavior.

Tests:
- TestHalfCloseWriteOrClose_PrefersCloseWriteWhenSupported pins
  the type-assertion path against a real net.TCPConn pair.
- TestHalfCloseWriteOrClose_FallsBackToCloseOnPlainConn pins the
  fallback so a future transport without CloseWrite doesn't leak.
- TestForwardRawTCP_DrainsBufferedBytesAfterLocalEOF is the e2e
  regression: 1 MB through forwardRawTCP, all bytes arrive, peer
  side validates byte-by-byte. Without the fix the immediate
  Close(remoteConn) races with the peer's reads and tail bytes
  drop.

Operator-impact: dmsg cat 50MB sends, skynet srv port-forwarded
bulk transfers, and iperf3 over skynet (which had been crashing
with rc=139 due to NULL-deref on premature EOF) should all
complete cleanly without partial-RST patterns.

* lint: errcheck on test cleanups + drive-by treeprobe linter fixes

#2735 CI surfaced two unrelated lint clusters:

(1) errcheck on forward_raw_tcp_drain_test.go: 5 Close/SetReadDeadline
    calls in cleanup paths. Added //nolint:errcheck. Errors at test
    teardown are not actionable.

(2) Drive-by on already-merged pkg/util/treeprobe (Gamma's #2734)
    that linter caught on this PR:
    - 'cancelled' misspell in a code comment (the field itself is
      correctly 'Canceled')
    - 'int32(a.runDone.TotalPinged)' was a no-op conversion since
      TotalPinged is already int32

No behavior change.

+++ pkg/util/treeprobe/aggregate.go
@@ -147,7 +147,7 @@ func (a *Aggregator) CacheHitRate() float64 {
 	if a.runDone == nil {
 		return 0
 	}
-	denom := int32(a.runDone.TotalPinged) + a.runDone.TotalSkippedCached
+	denom := a.runDone.TotalPinged + a.runDone.TotalSkippedCached
 	if denom == 0 {
 		return 0
 	}
